+++ commands/deploy.go
@@ -2,11 +2,15 @@ package commands
 
 import (
 	"bytes"
+	"context"
+	"crypto/md5"
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"github.com/sachamorard/swapper/response"
 	"github.com/sachamorard/swapper/utils"
 	"github.com/sachamorard/swapper/yaml"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"os"
@@ -15,6 +19,7 @@ import (
 	"time"
 
 	"github.com/docopt/docopt-go"
+	"cloud.google.com/go/storage"
 )
 
 var (
@@ -76,33 +81,82 @@ func Deploy(argv []string) response.Response {
 		return response.Fail(response.ErrorMessages["no_hash_field"])
 	}
 
-	if masterHostname == "localhost" {
-		masterHostname, _ = utils.GetHostname()
-	}
-	if masterHostname == "127.0.0.1" {
-		masterHostname, _ = utils.GetHostname()
-	}
-
-	i := strings.Index(masterHostname, ":")
-	if i == -1 {
-		masterHostname = masterHostname + ":1207"
-	}
-
-	conf := CurlRoot(masterHostname)
-	if len(conf.Yamls) == 0 {
-		return response.Fail(fmt.Sprintf(response.ErrorMessages["bad_master_addr"], masterHostname))
-	}
-
-	masterSplit := strings.Split(masterHostname, ":")
-	port := masterSplit[1]
-
 	fileInfo, _ := os.Stat(file)
 	var valid = regexp.MustCompile(`\.yml$`)
 	if valid.MatchString(fileInfo.Name()) == false {
 		return response.Fail(response.ErrorMessages["yaml_name"])
 	}
 
-	return DeployFile(fileInfo.Name(), cleanYaml, port, yamlConf)
+	if yamlConf.Master.Driver == "local" {
+		if masterHostname == "localhost" {
+			masterHostname, _ = utils.GetHostname()
+		}
+		if masterHostname == "127.0.0.1" {
+			masterHostname, _ = utils.GetHostname()
+		}
+
+		i := strings.Index(masterHostname, ":")
+		if i == -1 {
+			masterHostname = masterHostname + ":1207"
+		}
+
+		conf := GetMastersConf(masterHostname)
+		if len(conf.Yamls) == 0 {
+			return response.Fail(fmt.Sprintf(response.ErrorMessages["bad_master_addr"], masterHostname))
+		}
+
+		masterSplit := strings.Split(masterHostname, ":")
+		port := masterSplit[1]
+
+		return DeployFile(fileInfo.Name(), cleanYaml, port, yamlConf)
+	}
+
+	hasher := md5.New()
+	hasher.Write([]byte(cleanYaml))
+	hash := hex.EncodeToString(hasher.Sum(nil))
+	cleanYaml = cleanYaml+"hash: "+hash
+
+	if yamlConf.Master.Driver == "gcp" {
+		ctx := context.Background()
+
+		if yamlConf.Master.CredentialsFile != "" {
+			_ = os.Setenv("GOOGLE_APPLICATION_CREDENTIALS", yamlConf.Master.CredentialsFile)
+		}
+
+		// Creates a client.
+		client, errClient := storage.NewClient(ctx)
+		if errClient != nil {
+			return response.Fail(fmt.Sprintf("Failed to create client: %v", errClient))
+		}
+
+		// Sets the name for the new bucket.
+		bucketName := "swapper-master-"+yamlConf.Master.ProjectId
+
+		// Creates a Bucket instance if not exists
+		bucket := client.Bucket(bucketName)
+		_, err = bucket.Attrs(ctx)
+		if err != nil {
+			// Creates the new bucket.
+			if err := bucket.Create(ctx, yamlConf.Master.ProjectId, nil); err != nil {
+				msg := fmt.Sprintf("Deployment failed\nFailed to create bucket: %v", err)
+				_ = utils.SlackSendError(msg, yamlConf)
+				return response.Fail(msg)
+			}
+			fmt.Printf("Bucket %v created.\n", bucketName)
+		}
+
+		if err := gcsWrite(cleanYaml, client, bucketName, fileInfo.Name()); err != nil {
+			msg := fmt.Sprintf("Deployment failed\nCannot write object: %v", err)
+			_ = utils.SlackSendError(msg, yamlConf)
+			return response.Fail(msg)
+		}
+
+		_ = utils.SlackSendSuccess("Deployment succeed", yamlConf)
+		nodeInstruction := "To start a node, execute:\nswapper node start --join gs://"+bucketName+" --apply "+fileInfo.Name()
+		return response.Success("\n>> Deployment succeed\n"+nodeInstruction+"\n")
+	}
+
+	return response.Fail("")
 }
 
 func DeployFile(filename string, cleanYaml string, port string, yamlConf yaml.YamlConf) response.Response {
@@ -134,4 +188,17 @@ func DeployReq(filename string, cleanYaml string, port string) (err error) {
 	} else {
 		return nil
 	}
+}
+
+func gcsWrite(swapperYml string, client *storage.Client, bucketName string, object string) error {
+	ctx := context.Background()
+	src := strings.NewReader(swapperYml)
+	wc := client.Bucket(bucketName).Object(object).NewWriter(ctx)
+	if _, err := io.Copy(wc, src); err != nil {
+		return err
+	}
+	if err := wc.Close(); err != nil {
+		return err
+	}
+	return nil
 }