+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.2
 * chore: manage dependencies with go modules
 * feat: allow multi yaml file deployment on masters
+* feat: deploy with gcs
 
 ## 1.0.1
 * feat: insert command instruction in your swapper.yml that will be executed by the node. Here for [exemple](https://github.com/SachaMorard/swapper/tree/master/doc/swapper.yml.examples/7.swapper.with.command.yml)